sems/backend/pyproject.toml
[project]
name = "sems"
version = "0.1.0"
description = "Document Tracking System for Certificate II Examination to digitize and manage examination score sheets with automated data extraction and processing"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.123.10",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.44",
    "pyzbar>=0.1.9",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "httpx>=0.27.0",
    "reductoai>=0.15.0",
    "pypdf2>=3.0.1",
    "jinja2>=3.1.6",
    "reportlab>=4.4.7",
    "python-barcode>=0.16.1",
    "weasyprint>=67.0",
    "cachetools>=6.2.4",
    "scipy>=1.16.3",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
]

[dependency-groups]
dev = [
  {include-group = "lint"},
  {include-group = "test"},
  "pre-commit>=4.5.0",
  "alembic>=1.17.2",
  "alembic-utils>=0.8.8",
]
lint = [
    "black>=25.11.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "mypy>=1.19.0",
    "ruff>=0.14.8",
]
test = [
    "coverage>=7.12.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
    "pytest-postgresql>=7.0.2",
]
[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]

[tool.alembic]
script_location = "alembic"

[tool.ruff]
target-version = "py310"
line-length = 120
exclude = [
"alembic",
  "tools",
  "test",
  "bkp"
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
    "T201",   # print statements are not allowed
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true

[tool.ruff.lint.per-file-ignores]
"**/__init__.py" = ["F401", "F403"]
"app/main.py" = ["F403", "E402", "ARG001"]

[tool.coverage.run]
source = ["app"]
dynamic_context = "test_function"

[tool.coverage.report]
show_missing = true
sort = "-Cover"

[tool.coverage.html]
show_contexts = true
